Load policy script from configuration
