+++ app/constitution/scitt.js
@@ -24,6 +24,8 @@ actions.set("set_scitt_configuration",
             checkType(alg, "string", `configuration.policy.accepted_did_issuers[${i}]`);
           }
         }
+
+        // TODO: Validate type of policy_script
       }
 
       checkType(args.configuration.authentication, "object?", "configuration.authentication");
